requirements.txt
autopep8==1.6.0
Pillow==9.2.0
pycodestyle==2.8.0
streamdeck==0.9.1
toml==0.10.2
